// src/types/api.types.ts
export type Product = {
  id: number;
  title: string;
  price: number;
  description: string;
  images: string;
  created_at: string;
  updated_at: string;
  realPrice:number;
  weight:number
};